+++ VVBasics/src/VVSpriteControl.h
@@ -17,7 +17,7 @@ int					_spriteControlCount;
 	BOOL					spritesNeedUpdate;
 	NSEvent					*lastMouseEvent;
 	NSColor					*clearColor;
-	int						mouseDownModifierFlags;
+	long					mouseDownModifierFlags;
 	BOOL					mouseIsDown;
 	NSView					*clickedSubview;	//	NOT RETAINED
 }
